+++ NEWS.md
@@ -1,3 +1,10 @@
+# datelife 0.6.2
+  - data: updated `opentree_chornograms` object. It now has 253 chronograms from Open Tree of Life
+  - functions:
+    - added functions `matrix_to_table` and matrices_to_table` that go from a matrix of patristic distances to a table of taxon name pairs and respective node ages.
+    - bug fix on `use_calibrations_bladj` that used element $present_calibrations instead of $matched_calibrations
+    - get_otol_chronogram is another name to call get_opentree_chronograms
+
 # datelife 0.6.1
   - Functions:
     - `get_otol_chronograms` was updated and renamed to `get_opentree_chronograms`
@@ -15,6 +22,7 @@
   - Package website with `pkgdown`
   - documentation: expanded for all functions
   - examples and function files are written to tempdir()
+  - function rename: `get_biggest_phylo` to `get_biggest_multiphylo`
 
 # datelife 0.5.0